requirements.txt
keras==2.15.0
xgboost==2.0.2
tensorflow==2.15.0
scikit-learn==1.3.2
scikit-image==0.22.0
matplotlib==3.8.0
numpy==1.24.3
opencv==4.6.0
albumentations==1.3.1
pandas==2.1.4
patchify==0.2.3
Pillow==10.0.1
scipy==1.11.4
skimage==0.0
tqdm==4.66.1
segmentation_models==1.0.1
torch==2.1.0
torchmetrics==1.2.1
torchsummary==1.5.1
segmentation-models-pytorch==0.3.3

